requirements.txt
streamlit==0.87.0
boto3==1.18.0
pillow==8.4.0
python-docx==0.8.11
python-dotenv